日常提交

+++ api/handler/post.go
@@ -12,7 +12,9 @@ import (
 )
 
 func AddPost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-	AuthToken( w, r,2)
+	if !AuthToken(w, r, 2) {
+		return
+	}
 	req, _ := ioutil.ReadAll(r.Body)
 	pbody := &def.Post{}
 
@@ -33,7 +35,9 @@ func AddPost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 }
 
 func UpdatePost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-	AuthToken( w, r,2)
+	if !AuthToken(w, r, 2) {
+		return
+	}
 	pid := p.ByName("id")
 	pint, _ := strconv.Atoi(pid)
 	req, _ := ioutil.ReadAll(r.Body)
@@ -54,7 +58,9 @@ func UpdatePost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 }
 
 func DeletePost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-	AuthToken( w, r,4)
+	if !AuthToken(w, r, 4) {
+		return
+	}
 	pid, _ := strconv.Atoi(p.ByName("id"))
 	err := db.DeletePost(pid)
 	if err != nil {
